+++ CHANGELOG.md
@@ -4,6 +4,10 @@ FEATURES
 
 * Adding Gemini extension. See [189](https://github.com/hashicorp/terraform-mcp-server/pull/189)
 
+IMPROVEMENTS
+
+* Adding support for searching Terraform Actions documentation
+
 FIXES
 
 * Minor fixes to example configuration for VS Code, Cursor, etc.